requirements.txt
Flask==0.10.1
Flask-Cors==2.1.0
gunicorn==19.3.0
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
numpy==1.9.2
pandas==0.16.2
python-dateutil==2.4.2
python-firebase==1.2
pytz==2015.4
requests==2.7.0
six==1.9.0
Werkzeug==0.10.4
